backend/requirements.txt
# Arxelos — Python Dependencies

# Core
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
jinja2>=3.1.4

# ML / DL — Model 1: Brain Tumor Classifier
tensorflow>=2.16.0
Pillow>=10.3.0
numpy>=1.26.0

# PyTorch — uncomment for Model 2 (Virtual Lesions)
# torch>=2.3.0
# torchvision>=0.18.0

# NLP / RAG (Model 3) — uncomment when building RAG pipeline
# langchain>=0.2.0
# langchain-openai>=0.1.0
# chromadb>=0.5.0
# sentence-transformers>=3.0.0

# MLOps — uncomment as CI/CD matures
# mlflow>=2.14.0
# dvc>=3.51.0

# Dev / Testing
pytest>=8.2.0
httpx>=0.27.0
ruff>=0.4.0
